Move matchers to separate modules
